Allow to use challenge dataset files in CLEF challenge
