+++ app/controllers/dataset_files_controller.rb
@@ -1,4 +1,6 @@
 class DatasetFilesController < ApplicationController
+  include Concerns::DatasetFiles
+
   before_action :authenticate_participant!
   before_action :set_dataset_file,
                 only: [:destroy, :edit, :update]
@@ -9,37 +11,12 @@ class DatasetFilesController < ApplicationController
   before_action :check_participation_terms
   before_action :set_s3_direct_post,
                 only: [:new, :create, :edit]
+  before_action :set_dataset_files, only: :index # from DatasetFilesConcern
+  before_action :set_dataset_folders, only: :index # from DatasetFilesConcern
 
   def index
-    @dataset_files = policy_scope(DatasetFile).where(challenge_id: @challenge.id)
-    @dataset_files = @dataset_files.map do |dataset_file|
-      next dataset_file if dataset_file.hosting_location != 'Own S3'
-
-      result = Rails.cache.fetch(dataset_file_cache_key(dataset_file), expires_in: 5.minutes) do
-        Aws::FetchDatasetFileService.new(dataset_file: dataset_file).call
-      end
-
-      if result.success?
-        dataset_file.external_url       = result.value[:url]
-        dataset_file.external_file_size = result.value[:size]
-      else
-        dataset_file.error_message = result.value
-      end
-
-      dataset_file
-    end
-
-    @dataset_folders = policy_scope(DatasetFolder).where(challenge_id: @challenge.id)
-    @dataset_folders.each do |dataset_folder|
-      result = Rails.cache.fetch(dataset_folder_cache_key(dataset_folder), expires_in: 5.minutes) do
-        Aws::FetchDatasetFolderService.new(dataset_folder: dataset_folder).call
-      end
-
-      if result.success?
-        dataset_folder.dataset_files = result.value
-      else
-        dataset_folder.error_message = result.value
-      end
+    if @challenge.clef_task&.use_challenge_dataset_files?
+      redirect_to clef_task_task_dataset_files_path(@challenge.clef_task, challenge_id: @challenge.id)
     end
   end
 
@@ -154,13 +131,5 @@ class DatasetFilesController < ApplicationController
         key:                   "dataset_files/challenge_#{@challenge.id}/#{SecureRandom.uuid}_${filename}",
         success_action_status: '201',
         acl:                   'private')
-  end
-
-  def dataset_folder_cache_key(dataset_folder)
-    "aws-dataset-folder/#{dataset_folder.id}-#{dataset_folder.updated_at.to_i}"
-  end
-
-  def dataset_file_cache_key(dataset_file)
-    "aws-dataset-file/#{dataset_file.id}-#{dataset_file.updated_at.to_i}"
   end
 end
