+++ app/views/task_dataset_files/index.html.erb
@@ -1,17 +1,27 @@
 <%= render partial: 'shared/challenges/masthead', locals: { challenge: @challenge, challenge_rounds: @challenge_rounds, vote: @vote, follow: @follow } %>
 <%= render partial: 'challenges/show/subnav', locals: { challenge: @challenge } %>
 
-<!-- section -->
-<section class="section-p-sm">
-  <div class="container-fluid">
-    <div class="row">
-      <div class="col-12">
-        <ul class="list-group list-group-flush-px-0 list-group-dataset">
-          <h4><%= @clef_task.task %> dataset files</h4>
-          <hr/>
-          <%= render_challenge_resources(@current_participant_clef_task, @clef_task, @challenge) %>
-        </ul>
+<% if @clef_task.use_challenge_dataset_files? && participant_clef_task_status(@current_participant_clef_task) == 'registered' %>
+  <%= render partial: 'dataset_files/dataset_files' %>
+<% else %>
+  <!-- section -->
+  <section class="section-p-sm">
+    <div class="container-fluid">
+      <div class="row">
+        <div class="col-12">
+          <ul class="list-group list-group-flush-px-0 list-group-dataset">
+            <% if @clef_task.use_challenge_dataset_files? %>
+              <div class="col-md-8 col-lg-7">
+                <h4>Resources</h4>
+              </div>
+            <% else %>
+              <h4><%= @clef_task.task %> dataset files</h4>
+            <% end %>
+            <hr/>
+            <%= render_challenge_resources(@current_participant_clef_task, @clef_task, @challenge) %>
+          </ul>
+        </div>
       </div>
     </div>
-  </div>
-</section>
+  </section>
+<% end %>
