+++ db/migrate/20200603151708_add_use_challenge_dataset_files_to_clef_tasks.rb
@@ -1,0 +1,5 @@
+class AddUseChallengeDatasetFilesToClefTasks < ActiveRecord::Migration[5.2]
+  def change
+    add_column :clef_tasks, :use_challenge_dataset_files, :boolean, null: false, default: false
+  end
+end
